scaling potentiometer with getXRange

+++ include/Potentiometer.h
@@ -4,16 +4,16 @@
 
 class Potentiometer {
 public:
-    Potentiometer(uint8_t pin, uint16_t minOutput, uint16_t maxOutput);
+    Potentiometer(uint8_t pin, int minOutput, int maxOutput);
     void init();
     void readValue();
-    void setOutputRange(uint16_t minOutput, uint16_t maxOutput);
-    uint16_t getScaledValue();
+    void setOutputRange(int minOutput, int maxOutput);
+    int getScaledValue();
 private:
     uint8_t m_potPin;
     uint16_t m_potValueRaw;
-    uint16_t m_minOutput;
-    uint16_t m_maxOutput;
+    int m_minOutput;
+    int m_maxOutput;
 };
